next small refactoring around common.inc.php
